refactor: KkogKkogItem으로 모달 로직 응집
